Split and rename TournamentModDisplay component
